+++ HISTORY.rst
@@ -21,6 +21,7 @@ latest (unreleased)
 
 **Features and Improvements**
 * Add carrier accounts on partner
+* Add secondary vendor field on purchase order
 * Install module l10n_ch_import_cresus
 * Install module stock_picking_invoice_link
 * Activate the Drop shipping settings